+++ lib/restart-trigger.ts
@@ -1,2 +1,2 @@
-// restart triggered at 2026-03-11T10:09:56.390Z
-export const RESTART_TS = 1773223796390;
+// restart triggered at 2026-03-11T18:40:46.813Z
+export const RESTART_TS = 1773254446813;